refactor: authChecker 도메인 분리
# Conflicts:
#	src/main/java/com/followMe/delivery_server/delivery/domain/Delivery.java
