+++ src/main/java/com/followMe/delivery_server/delivery/domain/Shipment.java
@@ -8,6 +8,7 @@ import com.followMe.delivery_server.delivery.domain.exception.InvalidNodeInforma
 import com.followMe.delivery_server.delivery.domain.exception.InvalidShipmentStatusException;
 import com.followMe.delivery_server.delivery.domain.exception.NodeTypeMismatchException;
 import com.followMe.delivery_server.delivery.domain.exception.ShipmentNotFoundException;
+import com.followMe.delivery_server.delivery.domain.service.DeliveryPermissionChecker;
 import jakarta.persistence.*;
 import java.time.Instant;
 import java.util.ArrayList;
@@ -165,7 +166,10 @@ public class Shipment extends BaseAudit {
     transitionTo(ShipmentStatus.CANCELLED);
   }
 
-  public void updateShipmentStatus(ShipmentStatus status) {
+  public void updateShipmentStatus(
+      ShipmentStatus status, UserContext user, DeliveryPermissionChecker permissionChecker) {
+
+    permissionChecker.checkShipmentStatusUpdateAccess(user, this);
     switch (status) {
       case SHIPPED -> this.ship();
